+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Compatibility: BondageFurniture
+
 ## 1.0.21  (2026-05-17)
 
 * Compatibility: Make CostListCalculator.CostListAdjusted reentrant. (For compatibility with Nivarian Race)